Get tests compiling
