+++ geary-core/src/commonMain/kotlin/com/mineinabyss/geary/context/GearyContext.kt
@@ -1,7 +1,5 @@
 package com.mineinabyss.geary.context
 
-import com.mineinabyss.geary.systems.QueryContext
-
 public interface GearyContext :
     EngineContext,
     FormatsContext,
