+++ geary-core/src/commonMain/kotlin/com/mineinabyss/geary/context/QueryContext.kt
@@ -1,0 +1,7 @@
+package com.mineinabyss.geary.context
+
+import com.mineinabyss.geary.systems.QueryManager
+
+public interface QueryContext {
+    public val queryManager: QueryManager
+}
