+++ geary-core/src/commonMain/kotlin/com/mineinabyss/geary/helpers/GearyEntityWithOperator.kt
